Move state from GeneralForm to NewMovie
