+++ src/components/admin-panel/new-movie/NewMovie.tsx
@@ -3,8 +3,40 @@ import { faXmark } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import GeneralForm from "./general-form/GeneralForm";
 import ControlButtonGroup from "./control-button-group/ControlButtonGroup";
+import { useState, useEffect } from "react";
+import ApiService from "../../../service/ApiService";
+import { GeneralFormData } from "../../../types/FormData";
+import { Genre } from "../../../types/Genre";
+import { DateRange, Range, RangeKeyDict } from "react-date-range";
+import { SelectOptionType } from "../../../types/SelectOptionType";
 
 export default function NewMovie() {
+    let [genreOptions, setGenreOptions] = useState<SelectOptionType[]>();
+    let [calendarState, setCalendarState] = useState<Range[]>([{ startDate: new Date(), endDate: new Date(), key: 'selection' }]);
+    let [isDatePickerOpened, setIsDatePickerOpened] = useState(false);
+    let [formattedDateRange, setFormattedDateRange] = useState("");
+
+    let [formData, setFormData] = useState<GeneralFormData>({
+        title: "",
+        language: "",
+        startDate: "",
+        endDate: "",
+        director: "",
+        pgRating: "",
+        duration: "",
+        genre: [],
+        trailer: "",
+        synopsis: ""
+    });
+
+    useEffect(() => {
+        ApiService.get<Genre[]>("/genres")
+            .then(genresResponse => {
+                const genreOptions = genresResponse.map(genre => ({ value: genre.id, label: genre.name }))
+                setGenreOptions(genreOptions);
+            })
+            .catch(error => console.error("Error fetching data:", error));
+    }, [])
     return (
         <div className="add-movie-container">
             <div className="add-movie-heading">
